adversarial_fl_framework/requirements.txt
absl-py==0.7.1
astor==0.8.0
certifi==2019.6.16
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
#mkl-fft==1.0.12
#mkl-random==1.0.2
protobuf==3.8.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
tb-nightly==1.14.0a20190603
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
Werkzeug==0.15.4
wincertstore==0.2
wrapt==1.11.2
matplotlib==3.1.0
ConfigArgParse==0.15.1
tensorboardX==1.9
scipy
scikit-learn
orderedset